♻️ :refactor:z-index 수정 및 홈 메인 컴포넌트 분리
