+++ src/components/Home/CategoryMain/CategoryMain.module.scss
@@ -1,0 +1,98 @@
+@import '/src/styles/main.scss';
+
+.product-info {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+  width: 100%;
+}
+
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 1100px;
+
+  @include tablet {
+    width: 619px;
+    padding: 0 20px;
+  }
+
+  @include mobile {
+    width: 375px;
+  }
+}
+
+.title {
+  font-size: 20px;
+  font-weight: 300;
+}
+
+.title-point {
+  color: transparent;
+  background: $main_gradient;
+  background-clip: text;
+  font-weight: 600;
+}
+
+.itemList {
+  width: 1100px;
+
+  @include tablet {
+    width: 550px;
+  }
+
+  @include mobile {
+    width: 320px;
+  }
+}
+
+.item {
+  display: flex !important;
+  width: 250px !important;
+
+  @include tablet {
+    width: 190px !important;
+  }
+
+  @include mobile {
+    width: 150px !important;
+  }
+}
+
+.all-product {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin-top: 100px;
+  width: 1100px;
+
+  @include tablet {
+    width: 619px;
+    align-items: center;
+    margin-top: 50px;
+  }
+
+  @include tablet {
+    width: 375px;
+  }
+}
+
+.product-list-grid {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 29px;
+
+  @include tablet {
+    width: 590px;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 15px;
+  }
+
+  @include mobile {
+    width: 320px;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 20px;
+  }
+}
