+++ src/components/NavigationBar/NavigationBar.module.scss
@@ -21,7 +21,7 @@
   position: fixed !important;
   top: 0 !important;
   left: 0 !important;
-  z-index: 100 !important;
+  z-index: 98 !important;
 }
 
 .container {
